+++ acccounts.txt
@@ -1,0 +1,11 @@
+Account 1
+
+0x739cbf06c58489f2bcdf977ee4bf511e9852f19d804c4385170cca807c5f240e
+
+Account 2
+
+0xc5183e0a1a37f4da5eb3f3d23e9427c3774f0ec79a8935f861d2726ae0db400f
+
+password:
+
+password
