+++ js/backup/server.js
@@ -6,6 +6,7 @@ const Web3= require('web3');
 const port=8000
 let playerTurn="";
 
+
 let game={
 player1Address:"",
 player2Address:"",
@@ -21,7 +22,8 @@ player1Ready:false,
 player2Ready:false,
 move:0,
 roundsBet:{},
-totalBet:0
+totalBet:0,
+playerTurn:""
 }
 
 let gameBoard;
@@ -41,6 +43,15 @@ let initializeGameBoard=function(){
         8: ' ',
         9: ' '
     };
+}
+
+function isPlayer1(data){
+    if(data==game.player1Address){
+    return true;
+    }
+    else{
+        return false;
+    }
 }
 
 io.on('connection', function(socket) {
@@ -77,16 +88,20 @@ io.on('connection', function(socket) {
                 game.player1Address=data.account;
                 game.totalRounds=data.rounds;
                 game.roundsBet=data.roundsBet;
-                game.player1BalanceToken+=parseInt(data.tokens);
-                if(game.player1BalanceToken<parseInt(data.total)){
-                    console.log("\n Insufficient balance\n");
-                    socket.emit("insufficientBalance",{"message":"you do not have enough tokens","balance":game.player1BalanceToken})
-                }
-                else{
-                    console.log("If failed, u have sufficient balance",game.player1BalanceToken,data.total);
-                    socket.join(data.account);
+                game.player1BalanceToken=parseInt(data.tokens);
+                game.totalBet=parseInt(data.total);
+                // if(game.player1BalanceToken<parseInt(data.total)){
+                //     console.log("\n Insufficient balance\n");
+                //     socket.emit("insufficientBalance",{"message":"you do not have enough tokens","balance":game.player1BalanceToken})
+                // }
+                // else{
+                    socket.join("Game_"+data.account);
+                    // socket.emit("gameCreated",game);
+                    // console.log(io.sockets.sockets);
+                    // console.log(io.sockets.adapter.rooms) //Returns {room_1_id: {}, room_2_id: {}}
+                    // console.log(io.sockets.server.eio.clients) //Return client sockets
                     socket.emit("gameCreated",game);
-                }
+            // }
                 
             }
         });
@@ -96,14 +111,12 @@ io.on('connection', function(socket) {
     });
 
      socket.on('joinGame', function (data) {
-        console.log('Join a game event',data);
-        
+        // socket.emit("gameCreated",game);
         let web3=createConnection();
         web3.eth.defaultAccount = data.account;
         var Contract = web3.eth.contract(contractSource.abi)
         contractInstance = Contract.at(data.gameAddress);
         let tokens=parseInt(data.tokens) * 10;
-        console.log("\n Trying to connect to join game at ",data.gameAddress);
         contractInstance.setupPlayer2({
                     value: web3.toWei(String(tokens),"finney"),
                     gas: 4e6
@@ -113,17 +126,19 @@ io.on('connection', function(socket) {
                             if(result.blockNumber != null) {
                                 console.log('Game ready');
                                 game.player2Address=data.account;
-                                game.player2BalanceToken+=parseInt(data.tokens);
-                                if(game.player2BalanceToken<parseInt(data.total)){
-                                    console.log("\n Insufficient balance\n")
-                                    socket.emit("insufficientBalance",{"message":"you do not have enough tokens","balance":game.player2BalanceToken})
-                                }
-                                else{
-                                    console.log("If failed, u have sufficient balance",game.player2BalanceToken,data.total);
-                                    socket.join(game.player1Address);
-                                    io.sockets.in(game.player1Address).emit('gameRoomFull',game);
+                                game.player2BalanceToken=parseInt(data.tokens);
+                                // if(game.player2BalanceToken<parseInt(data.total)){
+                                //     console.log("\n Insufficient balance\n")
+                                //     socket.emit("insufficientBalance",{"message":"you do not have enough tokens","balance":game.player2BalanceToken})
+                                // }
+                                // else{
+                                    socket.join("Game_"+game.player1Address);
+                                    game.playerTurn=game.player1Address;
+                                    io.sockets.in("Game_"+game.player1Address).emit('gameRoomFull',game);
+                                    // console.log(io.sockets.adapter.rooms) //Returns {room_1_id: {}, room_2_id: {}}
+                                    // console.log(io.sockets.server.eio.clients)
                                     clearInterval(interval)
-                                }
+                                // }
                                 
                             }
                         })
@@ -138,13 +153,15 @@ io.on('connection', function(socket) {
         console.log('Create a player account event');
         let conn=createConnection();
         let account=conn.personal.newAccount();
-        console.log(account)
         socket.emit("accountCreated",{"account":account});
      },function(err){
         console.log("error occured try again",err);
         socket.emit("error",{"error":"error creating new account"});
     });
-
+    socket.on('totalBet',function(){
+        console.log("totalbet is triggered");
+        socket.emit("bets",game);
+    })
     socket.on("playerReady",function(data){
         if(game.player1Address==data.player){
             game.player1Ready=true;
@@ -154,14 +171,52 @@ io.on('connection', function(socket) {
         }
         if(game.player2Ready && game.player1Ready){
             initializeGameBoard();
-            playerTurn=game.player1Address;
-            io.sockets.in(game.player1Address).emit('gameReady',{"game":game,"gameboard":gameBoard});
+            console.log("Both players are ready");
+            game.playerTurn=game.player1Address;
+            io.sockets.in("Game_"+game.player1Address).emit('gameReady',{"game":game,"gameboard":gameBoard,"turn":game.playerTurn});
         }
     })
 
     socket.on('sendMsg',function(data){
-        io.sockets.in(game.player1Address).emit('message',data);
+        io.sockets.in("Game_"+game.player1Address).emit('message',data);
     })
+
+    socket.on('gameInitialised',function(){
+        io.sockets.in("Game_"+game.player1Address).emit('message',{"game":game,"board":gameBoard,"message":"\n\nFirst time\n\n","turn":game.playerTurn});
+    })
+
+    socket.on('moveMade',function(data){
+        console.log("\n Sending a message from server on making a move \n")
+        console.log("\nMove Made by\n",game.playerTurn,"\n\n",data);
+        console.log(isPlayer1(data.client))
+        if(gameBoard[data.move]==" "){
+            if(isPlayer1(data.client)){
+                gameBoard[data.move]="X";
+                console.log("was Player 1 ");
+                game.playerTurn=game.player2Address;
+                io.in("Game_"+game.player1Address).emit('message',{"game":game,"board":gameBoard,"message":"\n\nData changed in game \n\n","turn":game.playerTurn});
+            }
+            else{
+                gameBoard[data.move]="O"
+                console.log("was Player 2 ");
+                game.playerTurn=game.player1Address;
+                io.in("Game_"+game.player1Address).emit('message',{"game":game,"board":gameBoard,"message":"\n\nData changed in game \n\n","turn":game.playerTurn});
+            }
+            console.log("=====================Game data inside if================ \n",game)
+            // socket.emit('makeAMove',{"game":game,"board":gameBoard,"message":"\n\nData changed in game \n\n","turn":game.playerTurn});
+            
+        }
+        else{
+            console.log("\Inside Else Invalid move");
+            // socket.emit('makeAMove',{"game":game,"board":gameBoard,"message":"\n\nNot changed\n\n","turn":game.playerTurn});
+            io.sockets.in("Game_"+game.player1Address).emit('makeAMove',{"game":game,"board":gameBoard,"message":"\n\nNot changed\n\n","turn":game.playerTurn});
+        }
+        
+       
+        
+
+    })
+    
      socket.on('loadFromPrivateKey', function (data) {
         let conn=createConnection();
         console.log(data)
